More project cleanup
