--- mr_joe_2_abigail/tempCodeRunnerFile.py
@@ -1,2 +1,0 @@
-
-            planet_functions.data(SCREEN, CLOCK, planets)
