--- old/planets_01
@@ -1,77 +1,0 @@
-import pygame
-import os
-import math
-from planets_helper_00 import Planet
-
-
-os.system('cls')
-pygame.init()
-
-WIDTH, HEIGHT = (1300, 900)
-SCREEN = pygame.display.set_mode((WIDTH, HEIGHT))
-Planet.WIDTH = WIDTH
-Planet.HEIGHT = HEIGHT
-
-
-program_title = "Newtonian Orbital Simulation of the Inner Planets"
-background = pygame.image.load("stars.jpg").convert()
-
-WHITE = (255, 255, 255)
-BLACK = (0, 0, 0)
-RED = (255, 0, 0)
-GREEN = (0, 255, 0)
-BLUE = (0, 0, 255)
-GRAY = (128, 128, 128)
-DARK_GRAY = (80, 80, 80)
-YELLOW = (255, 255, 0)
-
-
-CLOCK = pygame.time.Clock()
-
-# fonts for text
-FONT_CS_36 = pygame.font.SysFont("comicsans", 36)
-title_text = FONT_CS_36.render(program_title, True, WHITE)
-abigail_text = FONT_CS_36.render("By: Abigail M. Lightle", True, WHITE)
-
-
-# def __init__(self, name, x,y,relative_radius,color,mass, y_vel ):
-sun = Planet("Sun", 0, 0, 2.5, YELLOW, 1.98892E30, 0)
-mercury = Planet("Mercury", -0.387, 0, 0.376, DARK_GRAY, 3.30e23, 47000)
-venus = Planet("Venus", -0.723, 0, 0.949, WHITE, 4.87E24, 35020)
-earth = Planet("Earth", -1, 0, 1.0, BLUE, 5.97E24, 29783)
-mars = Planet("Mars", -1.524, 0, 0.533, RED, 6.42E23, 24077)
-planets = [sun, earth, venus, mars, mercury]
-
-
-fps = 60
-pygame.display.set_caption(
-    f"{program_title}      fps - {fps}      by Abigail M. Lightle      Science Fair 2023-2024")
-ticks = 0
-run = True
-#################  START GAME LOOP  ###############
-while run == True:
-    CLOCK.tick(fps)
-    ticks += 1
-
-    # BLIT Text to Screen
-    SCREEN.blit(background, (0, 0))
-    SCREEN.blit(title_text, (225, 5))
-    SCREEN.blit(abigail_text, (915, 840))
-
-    for event in pygame.event.get():
-        if event.type == pygame.QUIT:
-            run = False
-        if event.type == pygame.KEYDOWN:
-            if event.key == pygame.K_p:
-                pass
-
-    for planet in planets:
-        # planet.update_position(planets)
-        planet.draw(SCREEN)
-        
-
-    pygame.display.flip()
-
-
-#################  END GAME LOOP ###############
-pygame.quit()
