+++ planet_global_variables.py
